Added additional flags to Makevars to allow compilation on Mac
I assume this is similar to Wes' update for the windows Makevars. Added flags to solve two errors: one which shows that the `validate_dims` member cannot be found (-DUSE_STANC3) and another which shows that the lgamma functions from the boost(?) library cannot be found (-DUSE_DARWIN_C_SOURCE and -D_REENTRANT).